uploadVideoFile
eslint - cleaned  up code for consistency

uploadVideoFile  - Ability to upload video to the server, store it in a "./media/video/..", then download the uploaded video for consistency (to have all video types video/mp4), after download completion delete original video and create thumbnails.

if error accrues due to ffmpeg delete original video and any other file that has been created.

As of now it's unknown to me how to get upload % when uploading the video to the server